This platform provides various functionalities with one special assistance functionality i.e. help/support/ask query through which users can connect with an ease.
Education is like a basic need similar to food, clothing and house nowadays. Looking at a present world education field is progressing day-by-day one gain education to not to only gain knowledge about specific thing but also to become a skillful in that field. In the modern tech world,  skilled people are demanded who compete with the existing problem and futuristic real time problems as well. The goal of this work is to provide a platform for learners to seek knowledge and form a strong foundation in their journey of mastering their respective field because if your foundation is strong then the building will be built firmly.